revert and add build to lint
