+++ .github/workflows/samples-tools/README.md
@@ -1,0 +1,7 @@
+# Cloud Samples tools
+
+This is a collection of tools used for Cloud Samples maintenance and infrastructure.
+
+## Finding affected packages
+
+## Running on all packages
